% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd.R
\name{assign_detection_flags}
\alias{assign_detection_flags}
\title{Assign quality control flags to values less than or equal to the minimum level of quantification but greater than the MDL.}
\usage{
assign_detection_flags(data)
}
\arguments{
\item{data}{The "data" table produced by getCCALData() after duplicates have been removed with handle_duplicates().}
}
\value{
The input data with the addition of a new flag field.
The only flag included in this routine is the J-R flag.
Any other flags will need to be defined separately by the user.
}
\description{
Assign quality control flags to values less than or equal to the minimum level of quantification but greater than the MDL.
}
\examples{
\dontrun{
getCCALData(file_path)[[1]][[1]] \%>\%
  handle_duplicates() \%>\%
  assign_detection_flags()
}
}
