% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd.R
\name{assign_flags}
\alias{assign_flags}
\title{Assign quality control flags.}
\usage{
assign_flags(data)
}
\arguments{
\item{data}{The "data" table produced by getCCALData() after duplicates have been removed with handle_duplicates().}
}
\value{
The input data with the addition of a new flag field.
Flags included in this automated routine are NFNSU, KRMDL, J-R, and NFNSI (in that order of priority).
Any other flags will need to be defined separately by the user.
}
\description{
Assign quality control flags.
}
\examples{
\dontrun{
getCCALData(file_path)[[1]][[1]] \%>\%
  handle_duplicates() \%>\%
  assign_flags()
}
}
