% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd.R
\name{clean_and_flag}
\alias{clean_and_flag}
\title{Clean and flag CCAL data}
\usage{
clean_and_flag(file_path)
}
\arguments{
\item{file_path}{Path to .xlsx file delivered by CCAL.}
}
\value{
The data with duplicate observations removed and flags added
}
\description{
Clean and flag CCAL data
}
\examples{
add something here?!?!
}
