% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd.R
\name{format_equis_results}
\alias{format_equis_results}
\title{Create the results table of the EQuIS EDD.}
\usage{
format_equis_results(
  file_paths,
  limits = imdccal::detection_limits,
  qualifiers = imdccal::equis_qualifiers,
  concat = FALSE
)
}
\arguments{
\item{file_paths}{Path to .xlsx file delivered by CCAL.}

\item{limits}{Table with detection limits. By default, uses the version in the package. User-defined versions must have the same columns.}

\item{qualifiers}{Table with flags and their meanings. By default, uses the version in the package. User-defined versions must have the same columns.}

\item{concat}{If concat is set to TRUE, the output contains one table rather than one for every input file.
By default, concat is set to FALSE, so the output contains separate tables for each file.
If only one file path is supplied to the file_paths argument, this parameter does not affect the output.}
}
\value{
The data formatted for input into the EQuIS system as the "Results" table.
Note that after running this function, the user must still complete some of their own processing to prepare their data for EQuIS.
For example, users must define Activity_ID, assign any additional flags, and modify Result_Status depending on the outcome of their quality control process.
}
\description{
Create the results table of the EQuIS EDD.
}
\examples{
# Edit limits table to work with example data
limits <- imdccal::detection_limits |>
  dplyr::mutate(EndDate = dplyr::if_else(EndDate == "2024-12-31", lubridate::ymd("2099-12-31"), EndDate))

# Create results table
results <- format_equis_results(file_paths = use_example_data(file_names = "SPAC_080199.xlsx"),
                          limits = limits)
}
