% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data_params}
\alias{get_data_params}
\title{List every parameter that occurs in the data}
\usage{
get_data_params(ccal_data)
}
\arguments{
\item{ccal_data}{The list object returned by \code{read_ccal}}
}
\value{
A character vector of parameter codes
}
\description{
List every parameter that occurs in the data
}
\examples{
my_folder <- "ccal_results"
file_list <- list.files(my_folder, pattern = "*.xlsx$", full.names = TRUE)
all_ccal_data <- read_ccal(file_list)
parameters <- get_data_params(all_ccal_data)
}
