% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_questionable_params}
\alias{get_questionable_params}
\title{List every parameter name that occurs in the questionable results}
\usage{
get_questionable_params(ccal_data)
}
\arguments{
\item{ccal_data}{The list object returned by \code{read_ccal}}
}
\value{
A character vector of parameter names
}
\description{
List every parameter name that occurs in the questionable results
}
\examples{
my_folder <- "ccal_results"
file_list <- list.files(my_folder, pattern = "*.xlsx$", full.names = TRUE)
all_ccal_data <- read_ccal(file_list)
param_names <- get_questionable_params(all_ccal_data)
}
