% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd.R
\name{handle_duplicates}
\alias{handle_duplicates}
\title{Remove lab duplicates and other duplicate rows.}
\usage{
handle_duplicates(data)
}
\arguments{
\item{data}{The "data" table produced by getCCALData().}
}
\value{
The data after lab QC duplicates have been removed and any duplicate rows resulting from the presence of multiple flags for one measurement have been removed.
}
\description{
Remove lab duplicates and other duplicate rows.
}
\examples{
\dontrun{
getCCALData(file_path)[[1]][[1]] \%>\%
  handle_duplicates()
}
}
