% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listParamsInData}
\alias{listParamsInData}
\title{List every parameter that occurs in the data}
\usage{
listParamsInData(ccal_data)
}
\arguments{
\item{ccal_data}{The list object returned by \code{getCCALData}}
}
\value{
A character vector of parameter codes
}
\description{
List every parameter that occurs in the data
}
\examples{
my_folder <- "ccal_results"
file_list <- list.files(my_folder, pattern = "*.xlsx$", full.names = TRUE)
all_ccal_data <- getCCALData(file_list)
parameters <- listParamsInData(all_ccal_data)
}
