% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listParamsInQuestionableResults}
\alias{listParamsInQuestionableResults}
\title{List every parameter name that occurs in the questionable results}
\usage{
listParamsInQuestionableResults(ccal_data)
}
\arguments{
\item{ccal_data}{The list object returned by \code{getCCALData}}
}
\value{
A character vector of parameter names
}
\description{
List every parameter name that occurs in the questionable results
}
\examples{
my_folder <- "ccal_results"
file_list <- list.files(my_folder, pattern = "*.xlsx$", full.names = TRUE)
all_ccal_data <- getCCALData(file_list)
param_names <- listParamsInQuestionableResults(all_ccal_data)
}
