% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{machineReadableCCAL}
\alias{machineReadableCCAL}
\title{Wrangle CCAL data into a machine-readable format}
\usage{
machineReadableCCAL(
  files,
  format = c("xlsx", "csv"),
  destination_folder = "./",
  overwrite = FALSE
)
}
\arguments{
\item{files}{Path to .xlsx file delivered by CCAL. Use a character vector to specify multiple files.}

\item{format}{File format to export machine readable data to - either "xlsx" or "csv"}

\item{destination_folder}{Folder to save the data in. Defaults to current working directory. Folder must already exist.}

\item{overwrite}{Overwrite existing file (Defaults to \code{TRUE} as with \code{write.table})}
}
\value{
Invisibly returns a list containing the data that were written to file.
}
\description{
Takes data as delivered by CCAL, extracts it, and rewrites it to tabs in an xlsx file or csv files in a folder.
}
\examples{
\dontrun{
ccal_folder <- "ccal_results"
dest_folder <- "ccal_results/tidied"
file_list <- list.files(my_folder, pattern = "*.xlsx$", full.names = TRUE)
machineReadableCCAL(file_list, format = "xlsx", destination_folder = dest_folder)
machineReadableCCAL(file_list, format = "csv", destination_folder = dest_folder)
}
}
