% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qualifiers}
\alias{qualifiers}
\title{EQuIS Flags}
\format{
\subsection{\code{qualifiers}}{

A data frame with 231 rows and 2 columns:
\describe{
\item{lookup_code}{The flag raised to represent a given concern about data.}
\item{remark}{A description of the meaning of a given flag.}
}
}
}
\source{
\url{https://doimspp.sharepoint.com/:x:/r/sites/nps-nrss-wrdiv/_layouts/15/Doc.aspx?sourcedoc=\%7B897FC8B3-2F68-4353-BB79-C0FEE9C45991\%7D&file=NPS_EQuIS_WQX_Reference_Values.xlsx&action=default&mobileredirect=true}
}
\usage{
qualifiers
}
\description{
Definitions of the various flags allowed by EQuIS.
}
\keyword{datasets}
