% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{range_to_vector}
\alias{range_to_vector}
\title{Expand Number Range in Text to All Numbers in Range}
\usage{
range_to_vector(numbers, separator = " ")
}
\arguments{
\item{numbers}{Character vector containing a number range indicated by a dash (e.g. "123 - 456")}

\item{separator}{Character string to use as a separator between numbers}
}
\value{
The same character vector, but with number ranges removed and replaced with the numbers in the range explicitly listed.
}
\description{
Takes a character vector containing a number range (e.g. "11 - 14") and replaces the range with the numbers contained in the range ("11 12 13 14").
}
\examples{
range_to_vector("1 3 5-8 10")
}
