% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd.R
\name{remove_ccal_duplicates}
\alias{remove_ccal_duplicates}
\title{Remove lab duplicates and other duplicate rows.}
\usage{
remove_ccal_duplicates(data)
}
\arguments{
\item{data}{The "data" table produced by read_ccal().}
}
\value{
The data after lab QC duplicates have been removed and any duplicate rows resulting from the presence of multiple flags for one measurement have been removed.
}
\description{
Remove lab duplicates and other duplicate rows.
}
\examples{
tidy_ccal_no_dups <- read_ccal(use_example_data(file_names = "SPAC_080199.xlsx"))[[1]][[1]] \%>\%
  remove_ccal_duplicates()
}
