% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_example_data}
\alias{use_example_data}
\title{List the file name(s) or file paths of imdccal's example data}
\usage{
use_example_data(file_names = NULL)
}
\arguments{
\item{file_names}{}
}
\value{
If file_names is supplied, \code{use_example_data()} returns the full file path to the requested package data.
Otherwise, it returns the file names of example data.
}
\description{
List the file name(s) or file paths of imdccal's example data
}
\examples{
# List all example data file names
use_example_data()

# Find file path to one example data file
use_example_data(file_names = "SPAC_080199.xlsx")

# Find file path to all example data files
use_example_data(file_names = use_example_data())
}
