% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{write_ccal}
\alias{write_ccal}
\title{Write data to xlsx or csv file.}
\usage{
write_ccal(
  all_data,
  format = c("xlsx", "csv"),
  destination_folder,
  overwrite,
  suffix,
  num_tables
)
}
\arguments{
\item{all_data}{The data to write to file.}

\item{format}{File format to export machine readable data to - either "xlsx" or "csv"}

\item{destination_folder}{Folder to save the data in. Defaults to current working directory. Folder must already exist.}

\item{overwrite}{Overwrite existing file (Defaults to \code{TRUE} as with \code{write.table})}

\item{suffix}{Suffix to add to output file name.}

\item{num_tables}{Number of tables to write to file.}
}
\description{
Write data to xlsx or csv file.
}
\examples{
\dontrun{
# Create tidied CCAL data from demo data stored in the imdccal package
tidy_ccal <- read_ccal(use_example_data(file_names = "SPAC_080199.xlsx"))

# Write data stored in environment to file
write_ccal(all_data = tidy_ccal,
          format = "xlsx", # alternatively, "csv"
          destination_folder = "ccal_tidy", # must already exist
          overwrite = TRUE,
          suffix = "_tidy",
          num_tables = 4)
}
}
