% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd.R
\name{format_results}
\alias{format_results}
\title{Create the results table of the EQuIS EDD.}
\usage{
format_results(
  file_paths,
  limits = imdccal::limits,
  qualifiers = imdccal::qualifiers
)
}
\arguments{
\item{file_paths}{Path to .xlsx file delivered by CCAL.}

\item{limits}{Table with detection limits. By default, uses the version in the package. User-defined versions must have the same columns.}

\item{qualifiers}{Table with flags and their meanings. By default, uses the version in the package. User-defined versions must have the same columns.}
}
\value{
The data formatted for input into the EQuIS system as the "Results" table.
Note that after running this function, the user must still complete some of their own processing to prepare their data for EQuIS.
For example, users must define Activity_ID, assign any additional flags, and modify Result_Status depending on the outcome of their quality control process.
}
\description{
Create the results table of the EQuIS EDD.
}
\examples{
# Edit limits table to work with example data
limits <- imdccal::limits |>
  dplyr::mutate(EndDate = dplyr::if_else(EndDate == "2024-12-31", lubridate::ymd("2099-12-31"), EndDate))

# Create results table
results <- format_results(file_paths = system.file("extdata", "SPAC_080199.xlsx", package = "imdccal"),
                          limits = limits)
}
