% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd.R
\name{format_results}
\alias{format_results}
\title{Create the results table of the EQuIS EDD}
\usage{
format_results(file_path)
}
\arguments{
\item{file_path}{Path to .xlsx file delivered by CCAL.}
}
\value{
The data formatted for EQuIS
}
\description{
Create the results table of the EQuIS EDD
}
\examples{
add something here?!?!
}
