% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd.R
\name{format_results}
\alias{format_results}
\title{Create the results table of the EQuIS EDD.}
\usage{
format_results(file_paths)
}
\arguments{
\item{file_path}{Path to .xlsx file delivered by CCAL.}
}
\value{
The data formatted for input into the EQuIS system as the "Results" table.
Note that after running this function, the user must still complete some of their own processing to prepare their data for EQuIS.
For example, users must define Activity_ID, assign any additional flags, and modify Result_Status depending on the outcome of their quality control process.
}
\description{
Create the results table of the EQuIS EDD.
}
\examples{
\dontrun{
results <- format_results(file_path)
}
}
