% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{getCCALData}
\alias{getCCALData}
\title{Read CCAL Data and convert to machine-readable format.}
\usage{
getCCALData(files)
}
\arguments{
\item{files}{Path to .xlsx file delivered by CCAL. Use a character vector to specify multiple files.}
}
\value{
A nested list. Each list item corresponds to one input file and contains data frames for data and metadata.
}
\description{
Read CCAL Data and convert to machine-readable format.
}
\examples{
tidy_ccal <- getCCALData(system.file("extdata", "SPAC_080199.xlsx", package = "imdccal"))
}
