% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imdccal-package.R
\docType{package}
\name{imdccal-package}
\alias{imdccal}
\alias{imdccal-package}
\title{imdccal: Wrangle CCAL Data Into Machine Readable Format}
\description{
This package takes water quality data spreadsheets (.xlsx) as delivered by the CCAL lab and reads them into R, including all metadata provided. Data can be exported to machine readable CSV or XLSX files that can easily be analyzed, imported into database tables, or added to an NPS data package for publication.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nationalparkservice.github.io/imd-ccal/}
}

}
\keyword{internal}
