% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{write_data}
\alias{write_data}
\title{Write data to xlsx or csv file.}
\usage{
write_data(
  all_data,
  format = c("xlsx", "csv"),
  destination_folder,
  overwrite,
  suffix,
  num_tables
)
}
\arguments{
\item{all_data}{The data to write to file.}

\item{format}{File format to export machine readable data to - either "xlsx" or "csv"}

\item{destination_folder}{Folder to save the data in. Defaults to current working directory. Folder must already exist.}

\item{overwrite}{Overwrite existing file (Defaults to \code{TRUE} as with \code{write.table})}

\item{suffix}{Suffix to add to output file name.}

\item{num_tables}{Number of tables to write to file.}
}
\description{
Write data to xlsx or csv file.
}
\examples{
\dontrun{
# Create tidied CCAL data from demo data stored in the imdccal package
tidy_ccal <- getCCALData(system.file("extdata", "SPAC_080199.xlsx", package = "imdccal"))

# Write data stored in environment to file
write_data(all_data = tidy_ccal,
          format = "xlsx", # alternatively, "csv"
          destination_folder = "ccal_tidy", # must already exist
          overwrite = TRUE,
          suffix = "_tidy",
          num_tables = 4)
}
}
