% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd.R
\name{write_results}
\alias{write_results}
\title{Write the results table of the EQuIS EDD.}
\usage{
write_results(
  files,
  format = c("xlsx", "csv"),
  destination_folder = "./",
  overwrite = FALSE
)
}
\arguments{
\item{files}{Path to .xlsx file delivered by CCAL. Use a character vector to specify multiple files.}

\item{format}{File format to export machine readable data to - either "xlsx" or "csv"}

\item{destination_folder}{Folder to save the data in. Defaults to current working directory. Folder must already exist.}

\item{overwrite}{Overwrite existing file (Defaults to \code{TRUE} as with \code{write.table})}
}
\value{
Invisibly returns a list containing the data that were written to file.
}
\description{
Write the results table of the EQuIS EDD.
}
\examples{
\dontrun{
ccal_folder <- "ccal_results"
dest_folder <- "ccal_results/edd"
file_list <- list.files(ccal_folder, pattern = "*.xlsx$", full.names = TRUE)
write_results(file_list, format = "xlsx", destination_folder = dest_folder)
write_results(file_list, format = "csv", destination_folder = dest_folder)
}
}
